REF: move scale_tau and its helper functions to robust.scale
